Cloud cover parameter changes for GFSv17 (#1027)
There is a low cloud bias in the current version of the GFSv17. The formula used in the cloud cover calculation is based Xu/Randall (1996). To increase the cloud cover in the GFSv17, a different set of parameters is used in the formulation. The increase in the cloud cover will also help to reduce the downward SW at the surface.